backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
langfuse>=2.0.0
